Restructure system tests module
